New class: SettingsManager
For now, it only manages back button preference
