+++ app/src/main/java/ch/logixisland/anuto/engine/theme/ThemeManager.java
@@ -22,7 +22,6 @@ public class ThemeManager extends BaseAdapter {
 
     private static final String PREF_FILE = "theme.prefs";
     private static final String PREF_THEME = "themeId";
-    private static final String PREF_BACK = "backButtonMode";
     private static final String PREF_TRANSPARENT_TOWER_INFO = "transparentTowerInfoEnabled";
 
     private final Context mContext;
@@ -33,36 +32,6 @@ public class ThemeManager extends BaseAdapter {
     private List<ThemeListener> mListeners = new CopyOnWriteArrayList<>();
     private TowerInfoFragment mVisibleTowerInfo;
 
-    public enum BackButtonMode {
-        DISABLED("DISABLED"), ENABLED("ENABLED"), TWICE("TWICE");
-
-        private final String code;
-        private static final Map<String, BackButtonMode> valuesByCode;
-
-        static {
-            valuesByCode = new HashMap<>(values().length);
-            for (BackButtonMode mode : values()) {
-                valuesByCode.put(mode.code, mode);
-            }
-        }
-
-        BackButtonMode(String code) {
-            this.code = code;
-        }
-
-        public static BackButtonMode modeFromCode(String code) {
-            return valuesByCode.get(code);
-        }
-
-        public String getCode() {
-            return code;
-        }
-    }
-    private BackButtonMode mBackButtonMode = BackButtonMode.DISABLED;
-
-    private static final long BACK_TWICE_INTERVAL = 2000L;//ms
-    private long mLastBackButtonPress = System.currentTimeMillis() - BACK_TWICE_INTERVAL;
-
     private boolean mTransparentTowerInfoEnabled;
 
     public ThemeManager(Context context) {
@@ -71,12 +40,6 @@ public class ThemeManager extends BaseAdapter {
 
         initThemes();
         loadTheme();
-
-        String backModeCode = mPreferences.getString(PREF_BACK, BackButtonMode.TWICE.getCode());
-        mBackButtonMode = BackButtonMode.modeFromCode(backModeCode);
-        if (mBackButtonMode == null) {
-            mBackButtonMode = BackButtonMode.DISABLED;
-        }
 
         mTransparentTowerInfoEnabled = mPreferences.getBoolean(PREF_TRANSPARENT_TOWER_INFO, false);
     }
@@ -121,20 +84,6 @@ public class ThemeManager extends BaseAdapter {
 
     public void setTheme(int index){
         setTheme(mAvailableThemes.get(index));
-    }
-
-    public BackButtonMode getBackButtonMode() {
-        return mBackButtonMode;
-    }
-
-    public void setBackButtonMode(BackButtonMode mode) {
-        if (mBackButtonMode != mode) {
-            mBackButtonMode = mode;
-
-            SharedPreferences.Editor editor = mPreferences.edit();
-            editor.putString(PREF_BACK, mBackButtonMode.getCode());
-            editor.apply();
-        }
     }
 
     public boolean isTransparentTowerInfoEnabled() {
@@ -189,26 +138,6 @@ public class ThemeManager extends BaseAdapter {
 
     public void removeListener(ThemeListener listener) {
         mListeners.remove(listener);
-    }
-
-    /** Tell the ThemeManager that user wants to exit with back button. Depending on the current
-     * BackButtonMode, the ThemeManager will allow to exit or not.
-     *
-     * @return true if BackButtonMode is ENABLED or if it is TWICE and this is the second press.
-     * false if BackButtonMode is DISABLED or if it is TWICE and this is the first press.
-     */
-    public boolean backButtonPressed() {
-        long timeNow = System.currentTimeMillis();
-        if(mBackButtonMode == BackButtonMode.DISABLED){
-            return false;
-        }else if(mBackButtonMode == BackButtonMode.ENABLED){
-            return true;
-        }else if(mLastBackButtonPress + BACK_TWICE_INTERVAL > timeNow){
-            return true;
-        }else{
-            mLastBackButtonPress = timeNow;
-            return false;
-        }
     }
 
     @Override
